+++ libpod/container_internal_linux.go
@@ -220,55 +220,6 @@ func (c *Container) reloadNetwork() error {
 	return c.save()
 }
 
-func (c *Container) getUserOverrides() *lookup.Overrides {
-	var hasPasswdFile, hasGroupFile bool
-	overrides := lookup.Overrides{}
-	for _, m := range c.config.Spec.Mounts {
-		if m.Destination == "/etc/passwd" {
-			overrides.ContainerEtcPasswdPath = m.Source
-			hasPasswdFile = true
-		}
-		if m.Destination == "/etc/group" {
-			overrides.ContainerEtcGroupPath = m.Source
-			hasGroupFile = true
-		}
-		if m.Destination == "/etc" {
-			if !hasPasswdFile {
-				overrides.ContainerEtcPasswdPath = filepath.Join(m.Source, "passwd")
-			}
-			if !hasGroupFile {
-				overrides.ContainerEtcGroupPath = filepath.Join(m.Source, "group")
-			}
-		}
-	}
-	if path, ok := c.state.BindMounts["/etc/passwd"]; ok {
-		overrides.ContainerEtcPasswdPath = path
-	}
-	return &overrides
-}
-
-func lookupHostUser(name string) (*runcuser.ExecUser, error) {
-	var execUser runcuser.ExecUser
-	// Look up User on host
-	u, err := util.LookupUser(name)
-	if err != nil {
-		return &execUser, err
-	}
-	uid, err := strconv.ParseUint(u.Uid, 8, 32)
-	if err != nil {
-		return &execUser, err
-	}
-
-	gid, err := strconv.ParseUint(u.Gid, 8, 32)
-	if err != nil {
-		return &execUser, err
-	}
-	execUser.Uid = int(uid)
-	execUser.Gid = int(gid)
-	execUser.Home = u.HomeDir
-	return &execUser, nil
-}
-
 // mountNotifySocket mounts the NOTIFY_SOCKET into the container if it's set
 // and if the sdnotify mode is set to container.  It also sets c.notifySocket
 // to avoid redundantly looking up the env variable.
